+++ crates/opentalk-controller-settings/src/settings_runtime/ws_rate_limit.rs
@@ -23,9 +23,9 @@ use crate::{
 #[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
 pub struct WebSocketRateLimit {
     /// The tokens that are added to the bucket per second
-    pub tokens_per_second: u64,
+    pub tokens_per_second: u16,
     /// The maximum amount of tokens that a token bucket can hold at a time
-    pub token_bucket_size: u64,
+    pub token_bucket_size: u16,
 }
 
 impl WebSocketRateLimit {
